rename(domain) : model 페키지명 domain으로 변경합니다
